+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- Exception during EXIF parsing for some Xiaomi images. Thanks to @mxsleo #92
 - (Pillow) `deepcopy` support for HeifImageFile and AvifImageFile
 - (macOS, arm64) `libde265`(HEIF decoder) now has the same version as in other builds(`1.0.8`->`1.0.11`)
 - (macOS, arm64) `libaom`(AVIF) now has the same version as in other builds(`3.4.0`->`3.6.0`)